Move FallbackFonts + measureFont to Font.ts
