+++ browser/src/Font.ts
@@ -1,5 +1,4 @@
-
-import { FallbackFonts } from "./Config"
+export const FallbackFonts = "Consolas,Monaco,Liberation Mono,DejaVu Sans Mono,Bitstream Vera Sans Mono,Courier New, monospace"
 
 export interface IFontMeasurement {
     width: number
