+++ browser/src/NeovimInstance.ts
@@ -5,7 +5,7 @@ import * as path from "path"
 import * as Q from "q"
 import * as Actions from "./actions"
 import * as Config from "./Config"
-import { measureFont } from "./measureFont"
+import { measureFont } from "./Font"
 import { Buffer, IBuffer } from "./neovim/Buffer"
 import { IWindow, Window } from "./neovim/Window"
 import * as Platform from "./Platform"
